+++ test/test_parsy.py
@@ -3,7 +3,8 @@ import unittest
 
 from parsy import test_char as parsy_test_char  # to stop pytest thinking this function is a test
 from parsy import (
-    ParseError, any_char, char_from, digit, generate, letter, line_info_at, regex, seq, string, string_from
+    ParseError, any_char, char_from, decimal_digit, digit, generate, letter, line_info_at, regex, seq, string,
+    string_from
 )
 
 
@@ -282,6 +283,11 @@ class TestParser(unittest.TestCase):
         self.assertEqual(any_char.parse("x"), "x")
         self.assertEqual(any_char.parse("\n"), "\n")
         self.assertRaises(ParseError, any_char.parse, "")
+
+    def test_decimal_digit(self):
+        self.assertEqual(decimal_digit.at_least(1).map(''.join).parse("9876543210"),
+                         "9876543210")
+        self.assertRaises(ParseError, decimal_digit.parse, "¹")
 
 
 class TestUtils(unittest.TestCase):