+++ docs/history.rst
@@ -11,6 +11,9 @@ History and release notes
   the parser was given a default description of the function name,
   which hides all useful internal info there might be.
 * Added :meth:`Parser.sep_by`
+* Added :func:`test`
+* Added :func:`char_from`
+* Added :func:`string_from`
 
 
 0.9.0 - 2017-09-28